requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
amqp==5.2.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
filelock==3.15.4
flatbuffers==24.3.25
flwr==1.10.0
fonttools==4.51.0
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
humanize==4.10.0
hydra-core==1.3.2
idna==3.7
iterators==0.0.2
jplephem==2.22
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.2.1
kiwisolver==1.4.5
kombu==5.3.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.0.8
namex==0.0.8
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
PyOpenGL==3.1.9
PyQt5
PyQt5-Qt5
PyQt5_sip
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.34.0
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.19.1
setuptools==69.5.1
sgp4==2.23
shellingham==1.5.4
six==1.16.0
skyfield==1.49
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
tk==0.1.0
tomli==2.0.1
tomli_w==1.0.0
tornado==6.4.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
vtk==9.3.1
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
